requirements.txt
scapy=2.6.1
rich=14.0.0
tqdm=4.67.1
pysmb=1.2.11